fix: optimisation simulator physics work commit
