+++ src/puzzle_gazebo/launch/launcher.launch
@@ -17,6 +17,7 @@
     <arg name="world_name" value="$(arg world_name)"/>
     <arg name="verbose" value="$(arg verbose)"/>
     <arg name="gui" default="$(arg gui)"/>
+    <arg name="physics" value="ode"/>
   </include>
 
   <!-- mavros -->
