+++ src/puzzle_gazebo/models/iris_puzzle/iris_puzzle.sdf.jinja
@@ -613,7 +613,7 @@
           </geometry>
         </visual>
         <sensor name="depth_camera" type="depth">
-          <update_rate>30</update_rate>
+          <update_rate>60</update_rate>
           <camera>
             <horizontal_fov>1.02974</horizontal_fov>
             <image>
@@ -638,7 +638,7 @@
             <cameraName>{{ namespase }}/camera</cameraName>
             <alwaysOn>1</alwaysOn>
             <useDepth16UC1Format>true</useDepth16UC1Format>
-            <updateRate>30</updateRate>
+            <updateRate>60</updateRate>
             <pointCloudCutoff>0.02</pointCloudCutoff>
             <pointCloudCutoffMax>18</pointCloudCutoffMax>
             <imageTopicName>rgb/image_raw</imageTopicName>
