+++ src/puzzle_gazebo/worlds/puzzle_pose_test.world
@@ -1,17 +1,41 @@
-<?xml version="1.0" ?>
+<?xml version="1.0"?>
 <sdf version="1.5">
   <world name="puzzle_world">
-
     <scene>
       <sky>
         <clouds>
-          <speed>12</speed>
+          <speed>12.0</speed>
         </clouds>
       </sky>
+      <shadows>true</shadows>
+      <grid>false</grid>
+      <origin_visual>false</origin_visual>
     </scene>
 
+    <physics type="ode">
+      <max_step_size>0.004</max_step_size>
+      <real_time_factor>1</real_time_factor>
+      <real_time_update_rate>250</real_time_update_rate>
+      <max_contacts>10</max_contacts>
+      <magnetic_field>6.0e-6 2.3e-5 -4.2e-5</magnetic_field>
+      <ode>
+        <solver>
+          <type>quick</type>
+          <iters>10</iters>
+          <sor>1.3</sor>
+          <use_dynamic_moi_rescaling>0</use_dynamic_moi_rescaling>
+        </solver>
+        <constraints>
+          <cfm>0</cfm>
+          <erp>0.2</erp>
+          <contact_max_correcting_vel>100</contact_max_correcting_vel>
+          <contact_surface_layer>0.001</contact_surface_layer>
+        </constraints>
+      </ode>
+    </physics>
+
     <gui>
-      <plugin name="ShootButton" filename="libGUIShootPlugin.so"/>
+      <plugin name="ShootButton" filename="libGUIShootPlugin.so" />
     </gui>
 
     <!-- A global light source -->
@@ -56,7 +80,6 @@
       <pose>3.5 -3.5 1.5 0 0 0</pose>
     </include>
 
-
     <include>
       <uri>model://iris</uri>
       <name>iris_222</name>
@@ -71,7 +94,7 @@
       <static>true</static>
     </include>
 
-     <include>
+    <include>
       <uri>model://big_box4</uri>
       <name>box_2</name>
       <pose>-3.5 3.5 1.5 0 0 0</pose>
@@ -84,7 +107,7 @@
       <static>true</static>
     </include>
 
-     <include>
+    <include>
       <uri>model://big_box4</uri>
       <name>box_3</name>
       <pose>-3.5 -3.5 1.5 0 0 0</pose>
